platform/x86/intel/vsec: Assign auxdev parent by argument
commit 6dfc2514acee37e30ce59f1f25b1f8f6aa7c1b08 upstream.

Instead of checking for a NULL parent argument in intel_vsec_add_aux() and
then assigning it to the probed device, remove this check and just pass the
device in the call. Since this function is exported, return -EINVAL if the
parent is not specified.

Intel-SIG: commit 6dfc2514acee platform/x86/intel/vsec: Assign auxdev parent by argument.
Backport intel tpmi base driver update for 6.6 from 6.10

Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com>
Link: https://lore.kernel.org/r/20231129222132.2331261-7-david.e.box@linux.intel.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
[ Yingbao Jia: amend commit log ]
Signed-off-by: Yingbao Jia <yingbao.jia@intel.com>