thermal/core: Check get_temp ops is present when registering a tz
commit 404f62cd6407f163e03cfaca97e27c1c4c62eb3c upstream.

Initially the check against the get_temp ops in the
thermal_zone_device_update() was put in there in order to catch
drivers not providing this method.

Instead of checking again and again the function if the ops exists in
the update function, let's do the check at registration time, so it is
checked one time and for all.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Krzysztof Sywula <krz.sywula@bytedance.com>